chore

+++ cote10830.cpp
@@ -5,12 +5,11 @@ class solver_10830{
 public:
 
     std::vector<std::vector<long long int>> matrixMul(std::vector<std::vector<long long int>> A, std::vector<std::vector<long long int>> B){
-
-        long long int n =A.size();
+        int n =A.size();
         std::vector<std::vector<long long int>> C(n, std::vector<long long int>(n));
-            for(long long int row =0 ; row < n ; ++row){
-                for(long long int col =0 ; col <n ; ++col){
-                    for(long long int prod =0 ; prod < n ; ++prod)
+            for(int row =0 ; row < n ; ++row){
+                for(int col =0 ; col <n ; ++col){
+                    for(int prod =0 ; prod < n ; ++prod)
                         C[row][col] += A[row][prod]*B[prod][col];
 
                     C[row][col] %= 1000;
@@ -20,7 +19,6 @@ public:
     }
 
     std::vector<std::vector<long long int>> matrixPow(std::vector<std::vector<long long int>> A, int pow){
-
         if (pow == 0 ) return ones;
         else if (pow == 1) return A;
         else if ( pow % 2 == 0){
@@ -31,7 +29,6 @@ public:
             std::vector<std::vector<long long int>> temp = matrixPow(A, pow-1);
             return matrixMul(temp, A);
         }
-
     }
 
     void insert(){
